+++ filip/clients/ngsi_v2/cb.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 
 import copy
 from copy import deepcopy
+from enum import Enum
 from math import inf
 from pkg_resources import parse_version
 from pydantic import PositiveInt, PositiveFloat, AnyHttpUrl
@@ -14,7 +15,7 @@ import re
 import requests
 from urllib.parse import urljoin
 import warnings
-from filip.clients.base_http_client import BaseHttpClient
+from filip.clients.base_http_client import BaseHttpClient, NgsiURLVersion
 from filip.config import settings
 from filip.models.base import FiwareHeader, PaginationMethod
 from filip.utils.simple_ql import QueryString
@@ -33,7 +34,6 @@ from filip.models.ngsi_v2.context import (
 from filip.models.ngsi_v2.base import AttrsFormat
 from filip.models.ngsi_v2.subscriptions import Subscription, Message
 from filip.models.ngsi_v2.registrations import Registration
-
 if TYPE_CHECKING:
     from filip.clients.ngsi_v2.iota import IoTAClient
 
@@ -71,6 +71,7 @@ class ContextBrokerClient(BaseHttpClient):
         """
         # set service url
         url = url or settings.CB_URL
+        self._url_version = NgsiURLVersion.v2_url.value
         super().__init__(
             url=url, session=session, fiware_header=fiware_header, **kwargs
         )
@@ -122,7 +123,7 @@ class ContextBrokerClient(BaseHttpClient):
             if res.ok:
                 items = res.json()
                 # do pagination
-                count = int(res.headers["Fiware-Total-Count"])
+                count = int(res.headers['Fiware-Total-Count'])
 
                 while len(items) < limit and len(items) < count:
                     # Establishing the offset from where entities are retrieved
@@ -167,7 +168,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
             Dict
         """
-        url = urljoin(self.base_url, "v2")
+        url = urljoin(self.base_url, self._url_version)
         try:
             res = self.get(url=url, headers=self.headers)
             if res.ok:
@@ -235,7 +236,7 @@ class ContextBrokerClient(BaseHttpClient):
                 the keyValues simplified entity representation, i.e.
                 ContextEntityKeyValues.
         """
-        url = urljoin(self.base_url, "v2/entities")
+        url = urljoin(self.base_url, f'{self._url_version}/entities')
         headers = self.headers.copy()
         params = {}
         options = []
@@ -345,7 +346,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, "v2/entities/")
+        url = urljoin(self.base_url, f'{self._url_version}/entities/')
         headers = self.headers.copy()
         params = {}
 
@@ -450,7 +451,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
             ContextEntity
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}")
+        url = urljoin(self.base_url, f'{self._url_version}/entities/{entity_id}')
         headers = self.headers.copy()
         params = {}
         if entity_type:
@@ -515,7 +516,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
             Dict
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs")
+        url = urljoin(self.base_url, f'{self._url_version}/entities/{entity_id}/attrs')
         headers = self.headers.copy()
         params = {}
         if entity_type:
@@ -681,7 +682,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
             None
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}")
+        url = urljoin(self.base_url, f'{self._url_version}/entities/{entity_id}')
         headers = self.headers.copy()
         if entity_type:
             params = {'type': entity_type}
@@ -809,7 +810,7 @@ class ContextBrokerClient(BaseHttpClient):
             None
 
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs")
+        url = urljoin(self.base_url, f'{self._url_version}/entities/{entity_id}/attrs')
         headers = self.headers.copy()
         params = {}
         if entity_type:
@@ -931,7 +932,7 @@ class ContextBrokerClient(BaseHttpClient):
             None
 
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs")
+        url = urljoin(self.base_url, f'{self._url_version}/entities/{entity_id}/attrs')
         headers = self.headers.copy()
         if entity_type:
             params = {"type": entity_type}
@@ -1029,7 +1030,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
             None
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs")
+        url = urljoin(self.base_url, f'{self._url_version}/entities/{entity_id}/attrs')
         headers = self.headers.copy()
         params = {}
         options = []
@@ -1097,7 +1098,8 @@ class ContextBrokerClient(BaseHttpClient):
             Error
 
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs/{attr_name}")
+        url = urljoin(self.base_url,
+                      f'{self._url_version}/entities/{entity_id}/attrs/{attr_name}')
         headers = self.headers.copy()
         params = {}
         if entity_type:
@@ -1166,7 +1168,8 @@ class ContextBrokerClient(BaseHttpClient):
             )
             attr_name = attr.name
 
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs/{attr_name}")
+        url = urljoin(self.base_url,
+                      f'{self._url_version}/entities/{entity_id}/attrs/{attr_name}')
         params = {}
         if entity_type:
             params.update({"type": entity_type})
@@ -1218,7 +1221,8 @@ class ContextBrokerClient(BaseHttpClient):
             Error
 
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs/{attr_name}")
+        url = urljoin(self.base_url,
+                      f'{self._url_version}/entities/{entity_id}/attrs/{attr_name}')
         headers = self.headers.copy()
         params = {}
         if entity_type:
@@ -1258,7 +1262,8 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs/{attr_name}/value")
+        url = urljoin(self.base_url,
+                      f'{self._url_version}/entities/{entity_id}/attrs/{attr_name}/value')
         headers = self.headers.copy()
         params = {}
         if entity_type:
@@ -1301,7 +1306,8 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs/{attr_name}/value")
+        url = urljoin(self.base_url,
+                      f'{self._url_version}/entities/{entity_id}/attrs/{attr_name}/value')
         headers = self.headers.copy()
         params = {}
         if entity_type:
@@ -1349,7 +1355,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, "v2/types")
+        url = urljoin(self.base_url, f'{self._url_version}/types')
         headers = self.headers.copy()
         params = {}
         if limit:
@@ -1378,7 +1384,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, f"v2/types/{entity_type}")
+        url = urljoin(self.base_url, f'{self._url_version}/types/{entity_type}')
         headers = self.headers.copy()
         params = {}
         try:
@@ -1401,7 +1407,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
             list of subscriptions
         """
-        url = urljoin(self.base_url, "v2/subscriptions/")
+        url = urljoin(self.base_url, f'{self._url_version}/subscriptions/')
         headers = self.headers.copy()
         params = {}
 
@@ -1513,7 +1519,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, f"v2/subscriptions/{subscription_id}")
+        url = urljoin(self.base_url, f'{self._url_version}/subscriptions/{subscription_id}')
         headers = self.headers.copy()
         try:
             res = self.get(url=url, headers=headers)
@@ -1559,7 +1565,7 @@ class ContextBrokerClient(BaseHttpClient):
                 DeprecationWarning,
             )
 
-        url = urljoin(self.base_url, f"v2/subscriptions/{subscription.id}")
+        url = urljoin(self.base_url, f'{self._url_version}/subscriptions/{subscription.id}')
         headers = self.headers.copy()
         headers.update({"Content-Type": "application/json"})
         try:
@@ -1586,7 +1592,8 @@ class ContextBrokerClient(BaseHttpClient):
         Args:
             subscription_id: id of the subscription
         """
-        url = urljoin(self.base_url, f"v2/subscriptions/{subscription_id}")
+        url = urljoin(self.base_url,
+                      f'{self._url_version}/subscriptions/{subscription_id}')
         headers = self.headers.copy()
         try:
             res = self.delete(url=url, headers=headers)
@@ -1611,7 +1618,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, "v2/registrations/")
+        url = urljoin(self.base_url, f'{self._url_version}/registrations/')
         headers = self.headers.copy()
         params = {}
 
@@ -1641,7 +1648,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, "v2/registrations")
+        url = urljoin(self.base_url, f'{self._url_version}/registrations')
         headers = self.headers.copy()
         headers.update({"Content-Type": "application/json"})
         try:
@@ -1669,7 +1676,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
             Registration
         """
-        url = urljoin(self.base_url, f"v2/registrations/{registration_id}")
+        url = urljoin(self.base_url, f'{self._url_version}/registrations/{registration_id}')
         headers = self.headers.copy()
         try:
             res = self.get(url=url, headers=headers)
@@ -1691,7 +1698,7 @@ class ContextBrokerClient(BaseHttpClient):
         Returns:
 
         """
-        url = urljoin(self.base_url, f"v2/registrations/{registration.id}")
+        url = urljoin(self.base_url, f'{self._url_version}/registrations/{registration.id}')
         headers = self.headers.copy()
         headers.update({"Content-Type": "application/json"})
         try:
@@ -1718,7 +1725,8 @@ class ContextBrokerClient(BaseHttpClient):
         Args:
             registration_id: id of the subscription
         """
-        url = urljoin(self.base_url, f"v2/registrations/{registration_id}")
+        url = urljoin(self.base_url,
+                      f'{self._url_version}/registrations/{registration_id}')
         headers = self.headers.copy()
         try:
             res = self.delete(url=url, headers=headers)
@@ -1783,7 +1791,7 @@ class ContextBrokerClient(BaseHttpClient):
 
         """
 
-        url = urljoin(self.base_url, "v2/op/update")
+        url = urljoin(self.base_url, f'{self._url_version}/op/update')
         headers = self.headers.copy()
         headers.update({"Content-Type": "application/json"})
         params = {}
@@ -1837,7 +1845,7 @@ class ContextBrokerClient(BaseHttpClient):
             follow the JSON entity representation format (described in the
             section "JSON Entity Representation").
         """
-        url = urljoin(self.base_url, "v2/op/query")
+        url = urljoin(self.base_url, f'{self._url_version}/op/query')
         headers = self.headers.copy()
         headers.update({"Content-Type": "application/json"})
         params = {"options": "count"}